_legacy/requirements.txt
numpy
opencv-python
matplotlib
pillow
unrealcv
Image